requirements.txt
async_timeout==4.0.2
discord.py==2.3.0
pyhedrals==0.2.0
yt_dlp==2023.7.6
PyNaCl==1.5.0
regex==2023.3.23
pycountry==22.3.5